Refactor utility resampling function
